backend/requirements_optimized.txt
# Core dependencies
Flask==3.0.1
Flask-Cors==4.0.0

# LLM inference (optimized)
llama_cpp_python==0.2.36

# PyTorch for GPU support (if needed)
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
safetensors>=0.3.0
accelerate>=0.20.0

# API and HTTP
aiohttp==3.9.2
httpx==0.26.0
httptools==0.6.1
uvicorn==0.27.0.post1

# Data processing
numpy==1.26.3
pandas==2.2.0
datasets>=2.10.0

# File handling
Werkzeug==3.0.1
filetype==1.2.0
pathlib==1.0.1

# Environment and configuration
python-dotenv==1.0.1
PyYAML==6.0.1

# Logging and monitoring
coloredlogs==15.0.1
humanfriendly==10.0

# System utilities
psutil==5.9.8
packaging==23.2
platformdirs==3.11.0

# Security
bcrypt==4.1.2

# JSON and data serialization
msgpack==1.0.7

# Async support
asyncio-throttle
aiosignal==1.3.1
async-timeout==4.0.3

# Optional: GPU acceleration
faiss-cpu==1.7.4

# Development and testing
pytest
pytest-asyncio

# Performance monitoring
memory-profiler

# Optional: Advanced text processing (only if needed)
# beautifulsoup4==4.12.3
# lxml==5.1.0
# nltk==3.8.1

# Remove these LangChain dependencies:
# langchain==0.1.4
# langchain-community==0.0.19
# langchain-core==0.1.22
# langsmith==0.0.87

# Remove other heavy dependencies if not needed:
# chromadb==0.4.22
# openai==1.11.1
# kubernetes==29.0.0
# opentelemetry-*